reorganize imports to avoid errors in other module
