+++ tomoscan/__init__.py
@@ -1,17 +1,17 @@
 __version__ = '0.1'
 
-from tomoscan.tomoscan import *
-from tomoscan.tomoscan_pso import *
-from tomoscan.tomoscan_helical import *
-from tomoscan.tomoscan_step import *
-from tomoscan.tomoscan_stream_pso import *
-from tomoscan.tomoscan_13bm_pso import *
-from tomoscan.tomoscan_13bm_mcs import *
-from tomoscan.tomoscan_2bm import *
-from tomoscan.tomoscan_2bm_step import *
-from tomoscan.tomoscan_6bm_step import *
-from tomoscan.tomoscan_32id import *
-from tomoscan.tomoscan_32id_step import *
-from tomoscan.tomoscan_stream_32id import *
-from tomoscan.tomoscan_stream_2bm import *
-from tomoscan.tomoscan_7bm import *
+# from tomoscan.tomoscan import *
+# from tomoscan.tomoscan_pso import *
+# from tomoscan.tomoscan_helical import *
+# from tomoscan.tomoscan_step import *
+# from tomoscan.tomoscan_stream_pso import *
+# from tomoscan.tomoscan_13bm_pso import *
+# from tomoscan.tomoscan_13bm_mcs import *
+# from tomoscan.tomoscan_2bm import *
+# from tomoscan.tomoscan_2bm_step import *
+# from tomoscan.tomoscan_6bm_step import *
+# from tomoscan.tomoscan_32id import *
+# from tomoscan.tomoscan_32id_step import *
+# from tomoscan.tomoscan_stream_32id import *
+# from tomoscan.tomoscan_stream_2bm import *
+# from tomoscan.tomoscan_7bm import *
