+++ tomoscan/tomoscan_stream_32id.py
@@ -32,7 +32,7 @@ import time
 import h5py 
 import numpy as np
 
-from tomoscan import TomoScanStreamPSO
+from tomoscan.tomoscan_stream_pso import TomoScanStreamPSO
 from tomoscan import log
 from tomoscan import util
 import threading
